Convert some doubles to long doubles and fix a bug
When setting the time rather than the number of iterations, the later
bins would be flushed empty after time up which would screw up the
results. Also I fixed a small bug in the command line progress bar.

+++ source/Corvette/CircularTime.hh
@@ -27,7 +27,7 @@ namespace SGF {
 */
 
 
-typedef double circulartime_t;
+typedef long double circulartime_t;
 
 class CircularTime {
 public:
